change balance to int64 and use money representation for operations, fix tests

+++ cmd/api/balance/messages.go
@@ -1,12 +1,12 @@
 package balance
 
 type BalanceMessage struct {
-	AccountID int     `json:"id"`
-	Amount    float64 `json:"amount"`
+	AccountID int   `json:"id"`
+	Amount    int64 `json:"amount"`
 }
 
 type TransferMessage struct {
-	FromID int     `json:"from"`
-	ToID   int     `json:"to"`
-	Amount float64 `json:"amount"`
+	FromID int   `json:"from"`
+	ToID   int   `json:"to"`
+	Amount int64 `json:"amount"`
 }
